+++ src/NeoVeldrid/TextureView.cs
@@ -55,6 +55,8 @@ namespace NeoVeldrid
         /// </summary>
         public abstract bool IsDisposed { get; }
 
+        bool BindableResource.IsBindable => !IsDisposed && !Target.IsDisposed;
+
         /// <summary>
         /// Frees unmanaged device resources controlled by this instance.
         /// </summary>